code_interpreter/requirements.txt
django
anthropic
PyPDF2
openpyxl
pandas
python-docx
